chore: move metadata to description tab in create flow (#2082)
